requirements.txt
pytest
customtkinter
time
unittest.mock
threading
pygame
pillow